Update AWG module
#### **The following changes are made in this update:**
##### **1) `_enable` parameter added to HDAWG's and UHFQA's AWG**
##### **2) `run` method of AWG Core updated:**
This method now uses the newly added `_enable` parameter and it enables
the AWG synchronously by default.
##### **3) `stop` method of AWG Core updated:**
This method now uses the newly added `_enable` parameter and it
disables the AWG synchronously by default.
##### **4) `wait_time` method of AWG Core updated:**
This method now checks if the AWG is running in continuous mode and
raises a `ToolkitError` if this is the case. It also raises a
`TimeoutError` if the AWG is not finished before timeout.
##### **5) Docstring improved**
##### **6) Tests for AWG Core improved**