Update `connection.py`
#### **The following changes are made in this update:**
##### **1) `update` method of `_scope_module` changed:**
This method is renamed to `update_device`, to make it consistent with 
other modules. This method is used to subscribe to the scope's data in 
the module. With this update, it will only call the `subscribe` method 
when the device using the scope mode is changed. Changes applied to 
`scope.py` as well.
##### **2) `device` attribute added to the scope module**
##### **3) `set` method of `DeviceConnection` class updated:**
`_commands_to_node` method is now also called before calling `sync_set`.
This is necessary because the device serial must be added to the node
address before calling this function so that we can set the nodes
synchronously using the following command, for example:
`uhfqa._set(scopes/0/enable, 1, sync=sync)`
##### **4) Exception type for invalid arguments changed :**
Instead of `ToolkitConnectionError`, a `TypeError` is raised for invalid
arguments.
##### **5) Fixed problems with local variables:**
Some local variables were being referenced before assignment, fixed the 
problem by initializing the local variables at the beginning of the 
function.
##### **6) `_get_value_from_streamingnode` changed to `staticmethod`**
##### **7) Docstrings improved**