Update PQSC drivers
#### **The following changes are made in this update:**
##### **1) `arm` method of PQSC updated:**
* `sync` argument is added and set to `True` by default. The `stop`
function and clear the register bank command are called synchronously
if this argument is set to `True`
* Type hints are added for the arguments `repetitions` and `holdoff`
##### **2) `run` method of PQSC updated:**
`sync` argument is added and set to `True` by default. The `_enable`
parameter is set to `True` synchronously if this argument
is set to `True`
##### **3) `arm_and_run` method added to PQSC:**
This method simply combines the methods `arm` and `run`
##### **4) `stop` method of PQSC updated:**
`sync` argument is added and set to `True` by default. The `_enable`
parameter is set to `False` synchronously if this argument
is set to `True`
##### **5) `wait_done` method of PQSC updated:**
This method now raises a `TimeoutError` if the PQSC is not finished
before timeout.
##### **6) Docstring improved**