Check sequence and trigger mode in `AWG` class
* The lines that check the sequence type and the trigger mode in
device specific `AWG` classes of `HDAWG` and `UHFQA` are exactly the
same. To avoid having duplicate code, move this check into the
`_apply_sequence_settings` method of the base `AWGCore` class and
inherit this method in the child classes before applying device
specific settings.
* If the sequence type or trigger mode is given as `None` as a string,
the enumeration for `None` is used.
* The return type hint `None` is also added to the method.