Update integration weight generation in `UHFQA`
#### **The following changes are made in this update:**
##### **1) `_demod_weights` updated:**
This method now generates both the real and imaginary part of the
integration weights. The real part corresponds to cosine and the
imaginary part corresponds to sine. Previously, the real part was sine
and the imaginary part was cosine. This method is now a `staticmethod`.
##### **2) `_set_int_weights` updated:**
The integration weights vector is obtained as a complex vector from
`_demod_weights`. Then it is seperated into real and imaginary parts
inside `_set_int_weights`. To send the integration weights to the
instrument `set_vector` is used instead of `set.`