Add new features to `SHFQA` driver
#### **The following changes are made in this update:**
##### **1) Add new parameter `sw_trigger`.**
##### **2) Add new modes to allowed trigger modes:**
* TriggerMode.RECEIVE_TRIGGER
* TriggerMode.ZSYNC_TRIGGER
At the moments, there are no particular settings applied in these
modes. However, they are still added since the Labber drivers of SHFQA
arm the Generator of SHFQA only if one of these trigger modes are
selected.
##### **3) `set_trigger_loopback` updated:**
The test source node is now set synchronously, similar to the API
examples in the repository
##### **4) New method `clear_trigger_loopback` added:**
This method stops the generation of internal trigger loopback pulses.
##### **5) `playback_delay` parameter added to `Generator`:**
##### **6) `Generator` parameter node paths updated:**
The `sequencer` word is removed from the node addresses in SHFQA. This
change is now reflected in zhinst-toolkit as well.
##### **7) `_apply_sequence_settings` moved the base `Generator`:**
Similar to HDAWG and UHFQA, this function is now called from the base
`Generator`. Device specific settings can still be added if desired.
##### **8) New parameters and methods added to `Sweeper`:**
* `oscillator_freq` and `integration_delay` parameters are added.
* `_trigger_source` attribute and `trigger_source` method are removed
and a new parameter with the name `trigger_source` is added.
* `output_freq` method sums the center frequency and offset frequency,
and it returns the output frequency.
* `get_result` method is renamed to `read` to make it consistent with
`Scope`