Fix some rare failure of wait_for_state_change.
In some rare cases, the function wait_for_state_change exited
early with a timeout error. That happened when a stale cached
value trigger an early detection of the state change. To avoid that,
a deep get is perfomed at the beginning and the logic is simplified.

Tests have been adapted to work with deep gets and to check the absence
of the previous faulty behaviour.