Fix wait_for_state change.
It was inadvertitly broken in 0e1bc9224096b14a7dcaea643d91f3d7b251c55d
The issue is that the zhinst core api sometimes return Python int (getInt),
and sometimes arrays of numpy int (get). As consequences, the result of a
comparison can be a numpy bool instead of bool, at that confused the
operator is, which check the equivalence of the objects, rather than
their value. Changed to a correct !=

Added a test case.